Fix issue #6310 (#6321)
check that cache directory exists before trying to list files there.

we try to remove all files in the cache directory that start with the function key, whether clear is called for self.persist
storage or not, to avoid leaving orphaned files in the cache directory. This leads to a bug, when cache directory is not exist, but we try to list all files there (to filter only those that belong to the corresponding function).

The solution is to check that directory exists before trying to filter corresponding files there.

+++ lib/streamlit/runtime/caching/storage/local_disk_cache_storage.py
@@ -193,9 +193,13 @@ class LocalDiskCacheStorage(CacheStorage):
         """Delete all keys for the current storage"""
         cache_dir = get_cache_folder_path()
 
-        for file_name in os.listdir(cache_dir):
-            if self._is_cache_file(file_name):
-                os.remove(os.path.join(cache_dir, file_name))
+        if os.path.isdir(cache_dir):
+            # We try to remove all files in the cache directory that start with
+            # the function key, whether `clear` called for `self.persist`
+            # storage or not, to avoid leaving orphaned files in the cache directory.
+            for file_name in os.listdir(cache_dir):
+                if self._is_cache_file(file_name):
+                    os.remove(os.path.join(cache_dir, file_name))
 
     def close(self) -> None:
         """Dummy implementation of close, we don't need to actually "close" anything"""
